fix models

+++ app/models.py
@@ -76,8 +76,8 @@ class ActorType(db.Model):
 
 class ReportViolations(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    reportId = db.Column(db.Integer(50), db.ForeignKey('report.id'))
-    violationId = db.Column(db.Integer(10), db.ForeignKey('violations.id'))
+    reportId = db.Column(db.Integer, db.ForeignKey('report.id'))
+    violationId = db.Column(db.Integer, db.ForeignKey('violations.id'))
     date_created = db.Column(db.DateTime, default=datetime.utcnow)
     
     def __repr__(self):
@@ -85,8 +85,8 @@ class ReportViolations(db.Model):
 
 class ViolationsSum(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    reportId = db.Column(db.Integer(50), nullable=False)
-    vioCategoryId = db.Column(db.Integer(10), nullable=False)
+    reportId = db.Column(db.Integer, nullable=False)
+    vioCategoryId = db.Column(db.Integer, nullable=False)
     summary = db.Column(db.Text)
     date_created = db.Column(db.DateTime, default=datetime.utcnow)
     
@@ -96,7 +96,7 @@ class ViolationsSum(db.Model):
 class Violations(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     violation = db.Column(db.String(512), nullable=False)
-    violationTypeId = db.Column(db.Integer(10))
+    violationTypeId = db.Column(db.Integer)
     reportViolations = db.relationship('ReportViolations', backref='violation')
 
     def __repr__(self):
@@ -117,7 +117,7 @@ class Orgs(db.Model):
     address2 = db.Column(db.String(256))
     city = db.Column(db.String(160))
     state = db.Column(db.String(4))
-    zipcode = db.Column(db.Integer(50))
+    zipcode = db.Column(db.Integer)
     phone = db.Column(db.String(50))
     orgEmail = db.Column(db.String(256))
     website = db.Column(db.String(256))
